requirements.txt
scapy
pysocks
stem